refactor: reorganize TypeScript configuration utilities into dedicated modules
