+++ src/tsconfig/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Provides functionality to read and validate TypeScript configuration files.
+ */
+
+export { parseConfig } from "./parser.ts";
+export { resolveProjectPath } from "./resolver.ts";
